+++ src/i18n.ts
@@ -1,6 +1,6 @@
 import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
-import { resources } from "lib/i18n/dictionary";
+import { resources } from "lib/i18n/dictionaries/dictionary";
 
 i18n.use(initReactI18next).init(
   {
