+++ src/lib/i18n/dictionaries/englishDictionary.ts
@@ -1,0 +1,254 @@
+import {
+  AchievementsDialog,
+  GeneralTerms,
+  GuideTerms,
+  Intro,
+  RulesTerms,
+  SeasonBannerOffer,
+  TranslationKeys,
+  WelcomeTerms,
+} from "./types";
+
+const generalTerms: Record<GeneralTerms, string> = {
+  featured: "Featured",
+  connecting: "Connecting",
+  loading: "Loading",
+  continue: "Continue",
+  readMore: "Read more",
+  close: "Close",
+  noThanks: "No thanks",
+};
+
+const welcomeTerms: Record<WelcomeTerms, string> = {
+  "welcome.otherWallets": "Other wallets",
+  "welcome.needHelp": "Need help?",
+  "welcome.createAccount": "Create account",
+  "welcome.login": "Login",
+  "welcome.signingIn": "Signing you in",
+  "welcome.signInMessage":
+    "Accept the signature request in your browser wallet to login.",
+};
+
+const rulesTerms: Record<RulesTerms, string> = {
+  rules: "Game Rules",
+  "rules.accounts": "1 account per player",
+  "rules.noBots": "No botting or automation",
+  "rules.game": "This is a game. Not a financial product.",
+  "rules.termsOfService": "Terms of Service",
+};
+
+const seasonBannerOffer: Record<SeasonBannerOffer, string> = {
+  "season.goodLuck": "Good luck in the season!",
+  "season.discount": "25% SFL discount on seasonal items",
+  "season.banner": "Seasonal Banner",
+  "season.wearableAirdrop": "Seasonal Wearable Airdrop",
+  "season.bonusTickets": "Bonus Seasonal Tickets",
+  "season.boostXP": "+10% EXP from food",
+  "season.exclusiveOffer": "Exclusive offer!",
+  "season.includes": "Includes:",
+  "season.limitedOffer": " Limited time only!",
+  "season.accessTo": "You have access to:",
+  "season.buyNow": "Buy Now",
+};
+
+const introTerms: Record<Intro, string> = {
+  "intro.one":
+    "Howdy there, partner! Welcome to Sunflower Land, the bountiful farming paradise!",
+  "intro.two":
+    "I'm Otis, an old Bumpkin farmer who's been tending to these lands for longer than I can remember. Truth be told, I could use a little help from a fresh face like you.",
+  "intro.three":
+    "You see, our little island has so much potential, and with your determination and hard work, we can transform it into a thriving empire!",
+  "intro.four":
+    "Looks like our little island is getting crowded. If we want to craft buildings and rare NFTs, we'll need more space.",
+  "intro.five":
+    "Let's first chop down these trees, gather some wood and expand the island.",
+};
+
+const achievementTerms: Record<AchievementsDialog, string> = {
+  "breadWinner.description": "Earn 0.001 SFL",
+  "breadWinner.one":
+    "Well, well, well, partner... It looks like you need some SFL!",
+  "breadWinner.two":
+    "In Sunflower Land, a healthy stash of SFL is the key to crafting tools, buildings and rare NFTs",
+  "breadWinner.three":
+    "The quickest way to earn SFL is by planting and selling crops.",
+
+  "sunSeeker.description": "Harvest Sunflower 100 times",
+  "cabbageKing.description": "Harvest Cabbage 200 times",
+  "jackOLantern.description": "Harvest Pumpkin 500 times",
+  "coolFlower.description": "Harvest Cauliflower 100 times",
+  "farmHand.description": "Harvest crops 10,000 times",
+  "beetrootBeast.description": "Harvest Beetroot 2,000 times",
+  "myLifeIsPotato.description": "Harvest Potato 5,000 times",
+  "rapidRadish.description": "Harvest Radish 200 times",
+  "twentyTwentyVision.description": "Harvest Carrot 10,000 times",
+  "stapleCrop.description": "Harvest Wheat 10,000 times",
+  "sunflowerSuperstar.description": "Harvest Sunflower 100,000 times",
+  "bumpkinBillionaire.description": "Earn 5,000 SFL",
+  "patientParsnips.description": "Harvest Parsnip 5,000 times",
+  "cropChampion.description": "Harvest 1 million crops",
+
+  "busyBumpkin.description": "Reach level 2",
+  "busyBumpkin.one":
+    "Howdy, my ambitious friend! To unlock new crops, expansions, buildings and much more you will need to level up.",
+  "busyBumpkin.two":
+    "Head over to the Fire Pit, cook up a delicious recipe and feed it to your Bumpkin.",
+
+  "kissTheCook.description": "Cook 20 meals",
+  "bakersDozen.description": "Bake 13 cakes",
+  "brilliantBumpkin.description": "Reach level 20",
+  "chefDeCuisine.description": "Cook 5,000 meals",
+
+  "scarecrowMaestro.description": "Craft a scarecrow and boost your crops",
+  "scarecrowMaestro.one":
+    "Howdy, partner! It is time you learn the art of crafting and boost your farming abilities",
+  "scarecrowMaestro.two":
+    "Travel to the Pumpkin Plaza, visit the Blacksmith and craft a Scarecrow.",
+
+  "bigSpender.description": "Spend 10 SFL",
+  "museum.description":
+    "Have 10 different kinds of rare items placed on your land",
+  "highRoller.description": "Spend 7,500 SFL",
+  "timbeerrr.description": "Chop 150 trees",
+  "craftmanship.description": "Craft 100 tools",
+  "driller.description": "Mine 50 stone rocks",
+  "ironEyes.description": "Mine 50 iron rocks",
+  "elDorado.description": "Mine 50 gold rocks",
+  "timeToChop.description": "Craft 500 axes",
+  "canary.description": "Mine 1,000 stone rocks",
+  "somethingShiny.description": "Mine 500 iron rocks",
+  "bumpkinChainsawAmateur.description": "Chop 5,000 trees",
+  "goldFever.description": "Mine 500 gold rocks",
+
+  // Explorer
+  "explorer.description": "Expand your land",
+  "explorer.one":
+    "Let's gather some wood by chopping down these trees and expand the island. Go ahead and figure out the best way to do it.",
+
+  "expansion.description": "Expand your land to new horizons.",
+
+  // Well of Prosperity
+  "wellOfProsperity.description": "Build a well",
+  "wellOfProsperity.one": "Well, well, well, what do we have here?",
+  "wellOfProsperity.two":
+    "It looks like your crops are thirsty. To support more crops you must first build a well.",
+
+  "contractor.description": "Have 10 buildings constructed on your land",
+  "fruitAficionado.description": "Harvest 50 fruit",
+  "fruitAficionado.one":
+    "Hey there, fruit gatherer! Fruits are nature's sweetest gifts, and they bring a burst of flavor to your farm.",
+  "fruitAficionado.two":
+    "By collecting different fruits, such as apples, oranges, and blueberries, you'll unlock unique recipes, boost your cooking skills, and create delightful treats",
+
+  "orangeSqueeze.description": "Harvest Orange 100 times",
+  "appleOfMyEye.description": "Harvest Apple 500 times",
+  "blueChip.description": "Harvest Blueberry 5,000 times",
+  "fruitPlatter.description": "Harvest 50,000 fruits",
+  "crowdFavourite.description": "Complete 100 deliveries",
+
+  "deliveryDynamo.description": "Complete 3 deliveries",
+  "deliveryDynamo.one":
+    "Howdy, reliable farmer! Bumpkins from all around need your help with deliveries.",
+  "deliveryDynamo.two":
+    "By completing deliveries, you'll make them happy and earn some fantastic SFL rewards in return ",
+
+  "seasonedFarmer.description": "Collect 50 Seasonal Resources",
+  "seasonedFarmer.one":
+    "Howdy, seasonal adventurer! Sunflower Land is known for its special seasons filled with unique items and surprises.",
+  "seasonedFarmer.two":
+    "By collecting Seasonal resources, you'll gain access to limited-time rewards, exclusive crafts, and rare treasures. It's like having a front-row ticket to the wonders of each season.",
+  "seasonedFarmer.three":
+    "So complete tasks, participate in events, and gather those Seasonal Tickets to enjoy the best that Sunflower Land has to offer!",
+  "treasureHunter.description": "Dig 10 holes",
+  "treasureHunter.one":
+    "Ahoy, treasure hunter! Sunflower Land is full of hidden treasures waiting to be discovered.",
+  "treasureHunter.two":
+    "Grab your shovel and head to Treasure Island, where you can dig for valuable items and rare surprises.",
+  "eggcellentCollection.description": "Collect 10 Eggs",
+  "eggcellentCollection.one":
+    "Howdy, egg collector! Chickens are wonderful farm companions that provide us with delicious eggs.",
+  "eggcellentCollection.two":
+    "By collecting eggs, you'll have a fresh supply of ingredients for cooking, and you'll also unlock special recipes and bonuses.",
+};
+
+const guideTerms: Record<GuideTerms, string> = {
+  "guide.intro":
+    "From humble beginnings to expert farming, this guide has got you covered!",
+  "gathering.description.one":
+    "To thrive in Sunflower Land, mastering the art of resource gathering is essential. Start by equipping the appropriate tools to collect different resources. Use the trusty Axe to chop down trees and acquire wood. To craft tools, visit the local workbench & exchange your SFL/resources for the desired tool.",
+  "gathering.description.two":
+    "As you progress and gather sufficient resources, you'll unlock the ability to expand your territory. Expanding your land opens up new horizons in Sunflower Land. Land expansions reveal a treasure trove of resources, including fertile soil for planting crops, majestic trees, valuable stone deposits, precious iron veins, shimmering gold deposits, delightful fruit patches and much more.",
+  "gathering.description.three":
+    "Remember, resource gathering and land expansion are the backbone of your farming journey. Embrace the challenges and rewards that come with each step, and watch your Sunflower Land flourish with bountiful resources and endless possibilities.",
+
+  "crops.description.one":
+    "In Sunflower Land, crops play a crucial role in your journey towards prosperity. By planting and harvesting crops, you can earn SFL (Sunflower Token) or utilize them to craft valuable recipes and items within the game.",
+  "crops.description.two":
+    "To grow crops, you need to purchase the respective seeds from the in-game shop. Each crop has a different growth time, ranging from just 1 minute for Sunflowers to 36 hours for Kale. Once the crops are fully grown, you can harvest them and reap the rewards.",
+  "crops.description.three":
+    "Remember, as you expand your land and progress in the game, more crops will become available, offering greater opportunities for earning SFL and exploring the vast potential of Sunflower Land's farming economy. So get your hands dirty, plant those seeds, and watch your crops flourish as you harvest your way to success!",
+
+  "building.description.one":
+    "Explore the diverse range of buildings available as you progress in Sunflower Land. From hen houses to workshops and beyond, each structure brings unique advantages to your farm. Take advantage of these buildings to streamline your farming operations, increase productivity, and unlock new possibilities. Plan your layout carefully and enjoy the rewards that come with constructing a thriving farm in Sunflower Land.",
+  "building.description.two":
+    "In Sunflower Land, buildings are the cornerstone of your farming journey. To access the buildings menu, click the Inventory icon and select the Buildings tab. Choose the desired structure and return to your farm screen. Find an open space, marked in green, and confirm the placement. Wait for the timer to complete, and your new building will be ready to use. Buildings provide various benefits and unlock exciting gameplay features. Strategically position them on your farm to maximize efficiency and watch as your farming empire grows and prospers.",
+
+  "cooking.description.one":
+    "Cooking allows you to nourish your Bumpkin and help them gain valuable experience points (XP). By utilizing the crops you've harvested, you can prepare delicious food at different buildings dedicated to cooking.",
+  "cooking.description.two":
+    "Starting with the Fire Pit, every farm has access to basic cooking facilities from the beginning. However, as you progress, you can unlock more advanced buildings such as the Kitchen, Bakery, Deli, and Smoothie Shack, each offering a wider variety of recipes and culinary delights.",
+  "cooking.description.three":
+    "To cook, simply select a building and choose a recipe you wish to prepare. The recipe will provide details about the required ingredients, the XP gained upon consumption, and the preparation time. After initiating the cooking process, keep an eye on the timer to know when the food will be ready to collect.",
+  "cooking.description.four":
+    "Once the food is ready, retrieve it from the building by clicking on it and moving it into your inventory. From there, you can interact with your Bumpkin NPC on the farm and feed them the prepared food, helping them gain XP and progress further in the game.",
+  "cooking.description.five":
+    "Experiment with different recipes, unlock new buildings, and discover the joy of cooking as you nurture your Bumpkin and embark on a delicious culinary adventure in Sunflower Land.",
+
+  "animals.description.one":
+    "Chickens in Sunflower Land are a delightful addition to your farm, serving as a source of eggs that can be used in various recipes and crafting. To start with chickens, you'll need to reach Bumpkin level 9 and build the Hen House. From there, you have the option to purchase chickens or place the ones you already have. Simply drag and drop them onto your farm, just like placing buildings. On a standard farm, every Hen House houses up to 10 chickens, and if you own the Chicken Coop SFT, this limit extends to 15.",
+  "animals.description.two":
+    "Each chicken has an indicator above its head, displaying its current mood or needs. This can range from being hungry, tired, happy, or ready to hatch. To keep your chickens content and productive, feed them by selecting wheat from your inventory and interacting with the chicken. Feeding initiates the egg timer, which takes 48 hours for the eggs to be ready to hatch. Once the eggs are ready, visit your farm, check the icon above each chicken, and interact with them to find out the type of egg that has hatched. Occasionally, you may even discover rare mutant chickens, which offer special boosts such as faster egg production, increased yield, or reduced food consumption.",
+  "animals.description.three":
+    "Nurturing your chickens and collecting their eggs adds a dynamic and rewarding element to your farm in Sunflower Land. Experiment with recipes, make use of the eggs in your crafting endeavors, and enjoy the surprises that come with rare mutant chickens. Build a thriving poultry operation and reap the benefits of your hard work as you embrace the charming world of chickens in Sunflower Land.",
+
+  "crafting.description.one":
+    "In Sunflower Land, crafting NFTs is a crucial aspect of boosting your farming output and accelerating your progress. These special items provide various bonuses, such as crop growth boosts, cooking enhancements, and resource boosts, which can greatly expedite your journey. By maximizing your SFL (Sunflower Token), you can craft tools, gather resources, and expand your land to further establish your farming empire.",
+  "crafting.description.two":
+    "To begin crafting items, we'll visit Igor, a skilled craftsman in Sunfloria. After hopping on the boat and arriving at Sunfloria, head to the top of the island to have a conversation with Igor. He is currently offering a Basic Scarecrow, which boosts the speed of Sunflowers, Potatoes, and Pumpkins. This is an excellent deal that requires exchanging your resources for the scarecrow. Once obtained, return to your main island and enter design mode by clicking on the white hand icon in the top right corner of the game.",
+  "crafting.description.three":
+    "In design mode, you can strategically place items and rearrange resources on your farm to optimize its layout and enhance its visual appeal. This step is crucial in maximizing the effectiveness of your crafted equipment. For example, place the Scarecrow over the plots you want to boost. Additionally, consider purchasing decorations to add charm and tidiness to your land.",
+  "crafting.description.four":
+    "By crafting equipment and placing it strategically, you can amplify your farming abilities, create an island home to be proud of, and accelerate your progress in Sunflower Land.",
+
+  "deliveries.description.one":
+    "Deliveries in Sunflower Land provide an exciting opportunity to help hungry Goblins and fellow Bumpkins while earning rewards. Every day you will be able to see all the orders you have by clicking on the delivery board on the bottom left of the screen. The orders have been placed by some local NPCs that can be found hanging around Pumpkin Plaza. To fulfill an order, you will need to take a boat ride to Pumpkin Plaza and look for the NPC expecting the delivery. Once you find them, click on them to deliver the order and receive your reward.",
+  "deliveries.description.two":
+    "As a new player, you start with three order slots, but as you expand your farm, you will unlock additional slots, allowing advanced players to take on more orders. New orders come in every 24 hours, offering a range of tasks from farming produce to cooking food and gathering resources. Completing orders will earn you milestone bonuses, including Block Bucks, SFL, delicious cakes, and other rewards. The reward system is based on the difficulty of the request, so consider prioritizing orders that offer greater rewards to maximize your gains. Keep an eye on the board and challenge yourself with a variety of orders, leveling up and unlocking new buildings as needed to fulfill more demanding requests.",
+
+  "scavenger.description.one":
+    "Scavenging in Sunflower Land offers exciting opportunities to uncover hidden treasures and gather valuable resources. The first aspect of scavenging is digging for treasure on Treasure Island, where you can become a pirate treasure hunter. By crafting a sand shovel and venturing to Treasure Island, you can dig in dark sandy areas to uncover a variety of treasures, including bounty, decorations, and even ancient SFTs with utility.",
+  "scavenger.description.two":
+    "Another form of scavenging involves gathering wild mushrooms that appear spontaneously on your farm and surrounding islands. These mushrooms can be collected for free and used in recipes, quests, and crafting items. Keep an eye out for these mushrooms, as they replenish every 16 hours, with a maximum limit of 5 mushrooms on your farm. If your land is full, mushrooms will appear on the surrounding islands, ensuring you don't miss out on these valuable resources.",
+
+  "fruit.description.one":
+    "Fruit plays a significant role in Sunflower Land as a valuable resource that can be sold for SFL or utilized in various recipes and crafting. Unlike crops, fruit patches have the unique ability to replenish multiple times after each harvest, providing a sustainable source of fruit for players.",
+  "fruit.description.two":
+    "To plant fruit, you'll need to acquire larger fruit patches, which become available on the 9-10th expansion of your farm.",
+  "fruit.description.three":
+    "By cultivating fruit and incorporating it into your farming strategies, you can maximize your profits, create delicious recipes, and unlock new possibilities in Sunflower Land.",
+
+  "seasons.description.one":
+    "Seasons in Sunflower Land bring excitement and freshness to the game, offering players new challenges and opportunities. With the introduction of each season, players can look forward to a variety of new craftable items, limited edition decorations, mutant animals, and rare treasures. These seasonal changes create a dynamic and evolving gameplay experience, encouraging players to adapt their strategies and explore new possibilities on their farms. Additionally, seasonal tickets add a strategic element to the game, as players must decide how to allocate their tickets wisely, whether it's collecting rare items, opting for higher supply decorations, or exchanging tickets for SFL. The seasonal mechanic keeps the game engaging and ensures that there's always something to look forward to in Sunflower Land.",
+  "seasons.description.two":
+    "The availability of seasonal items at the Goblin Blacksmith adds another layer of excitement. Players must gather the required resources and seasonal tickets to craft these limited-supply items, creating a sense of competition and urgency. Planning ahead and strategizing become crucial as players aim to secure their desired items before the supply runs out. Moreover, the option to swap seasonal tickets for SFL provides flexibility and allows players to make choices that align with their specific gameplay goals. With each season's unique offerings and the anticipation of surprise events, Sunflower Land keeps players engaged and entertained throughout the year, fostering a vibrant and ever-evolving farming experience.",
+};
+
+export const ENGLISH_TERMS: Record<TranslationKeys, string> = {
+  ...generalTerms,
+  ...welcomeTerms,
+  ...rulesTerms,
+  ...seasonBannerOffer,
+  ...introTerms,
+  ...achievementTerms,
+  ...guideTerms,
+};
