+++ src/lib/i18n/translate.ts
@@ -1,6 +1,6 @@
 import i18n from "src/i18n";
-import { TranslationKeys } from "./dictionary";
 import { TOptions } from "i18next";
+import { TranslationKeys } from "./dictionaries/types";
 
 export const translate = (key: TranslationKeys, options?: TOptions): string => {
   return i18n.t(key, options);
